This code uses the raw data to aggregate - from samples to strings. We have 28 strings with 5 non-independent samples on each string. So – we will focus on the following metrics: Abundance as MaxN: Species/group specific Max N per string (This uses the maximum count of each species across the samples within a string, recognizing that the 5 samples on a string are not independent observations). We have 2-3 strings per site – so we can get an average MaxN for a species/group in each site based on 2-3 data points.
Biomass:
Here we will follow a similar approach as with MaxN – considering the issue of double counting fish across samples from the same string.
So, we will use the maximum biomass for a species/group found across the 5 samples – and use as the biomass estimate for that species/group in that string. We then again have 2-3 biomass estimates per species/group for each site.
Diversity
This is straightforward enough – and will be the maximum number of different species/group across all samples in a string. We will then get diversity per string and can average it across 2-3 strings in each site.